Feat : Calcul_IQM.py overhaul
Implements the use of the extract sub-watersheds and individual A1, A2, A3 and F1 scripts in the Calcul_IQM script. Correction to the final IQM calculation that used 38 as a max score instead of 40. Updates the input params of some processes. Add some journal and progress bar updates to follow progression. Add try and except statement for better exception handling. Add a timing function to output the duration of each steps. Minor style and documentation improvements especially to segment the code into clear steps.

Close : Issue #35