Refractor : revision of A4 script
Correction to make the calculation more robust to disjoint lines that distort distance calculations. Add a header with the correct licence statement and standardised the formatting by using tabs for indents. Add try except statements to handle exceptions, output intermediate results to the final attribute table. Updated the shortHelpString to emphasise the use of the sinuosity index rather than the complexity index for calculating A4. Clean up the code and reduced verbosity.

Refs : Issue #38, #39, #14, #9 and #6