Perf : F3 performance refactoring
Modify the whole code to be more efficient and scalable for bigger watersheds. Leverage the speed of geometry objects and spatial indexes to speed the intersection with the riparian zone. Use a variable transects strategy to use a fixed number of transects on long streams and a given step for smaller ones. Account for both river banks interaction with the riparian zone. Fix the scoring method to fit the definition of the index. Implement a warning message to document if the river segment is < 2m (probably a clipping artefact from data preprocessing. Improve documentation for each step.

Refs : Issue #40