Fix : Projection and railway selection
Implement a reprojection into the project's CRS if it is not the same as the project. Removes the option to select abandoned railway since they are often not there in real life. Add the selection of cycleway. Update the shortHelpString to add the data source of the watershed area layer.